Replace old parser with RESP3 parser and remove hiredis
The hiredis parser does not support RESP3 nor will it support it any
time soon (at least unlikely). Therefore it is best to remove it.
